+++ index.js
@@ -44,7 +44,7 @@ bitcore.errors = require('./lib/errors');
 bitcore.Address = require('./lib/address');
 bitcore.Block = require('./lib/block');
 bitcore.MerkleBlock = require('./lib/block/merkleblock');
-bitcore.MnDiffList = require('./lib/mnlists/mnDiffList');
+bitcore.MnListDiff = require('./lib/mnlists/mnListDiff');
 bitcore.BlockHeader = require('./lib/block/blockheader');
 bitcore.HDPrivateKey = require('./lib/hdprivatekey.js');
 bitcore.HDPublicKey = require('./lib/hdpublickey.js');
